+++ src/enums/DrawMode.ts
@@ -1,0 +1,6 @@
+enum ShapeDrawMode {
+    Fill,
+    Stroke
+}
+
+export default ShapeDrawMode