+++ src/interfaces/jsons/JsonRelativeShape.ts
@@ -1,0 +1,5 @@
+import ISimplePoint from '../ISimplePoint'
+
+type JsonRelativeShape = ISimplePoint[]
+
+export default JsonRelativeShape