+++ src/objects/UIObject.ts
@@ -1,3 +1,5 @@
 import GameEntity from './GameEntity'
 
-export default class UIObject extends GameEntity {}
+export default class UIObject extends GameEntity {
+    
+}