+++ src/utils/static/Logging.ts
@@ -1,10 +1,10 @@
-import LoggingDebugLevel from '../../enums/LoggingDebugLevel'
+import LoggingLevel from '../../enums/LoggingLevel'
 
 export default class Logging {
-    static debugLevel: LoggingDebugLevel = LoggingDebugLevel.None
+    static debugLevel: LoggingLevel = LoggingLevel.None
     static isShowTrace: boolean = false
 
-    static message(cb: (...values: any[]) => void, needLevel: LoggingDebugLevel, ...values: any[]): void {
+    static message(cb: (...values: any[]) => void, needLevel: LoggingLevel, ...values: any[]): void {
         if(Logging.debugLevel >= needLevel) {
             cb(...values)
             if(Logging.isShowTrace) console.trace(cb.name)
@@ -12,34 +12,34 @@ export default class Logging {
     }
 
     static error(...values: any[]): void {
-        Logging.message(console.error, LoggingDebugLevel.Error, ...values)
+        Logging.message(console.error, LoggingLevel.Error, ...values)
     }
 
     static warn(...values: any[]): void {
-        Logging.message(console.warn, LoggingDebugLevel.Warn, ...values)
+        Logging.message(console.warn, LoggingLevel.Warn, ...values)
     }
 
     static debug(...values: any[]): void {
-        Logging.message(console.log, LoggingDebugLevel.Debug, ...values)
+        Logging.message(console.log, LoggingLevel.Debug, ...values)
     }
 
     static info(...values: any[]): void {
-        Logging.message(console.info, LoggingDebugLevel.Info, ...values)
+        Logging.message(console.info, LoggingLevel.Info, ...values)
     }
 
     static engineError(...values: any[]): void {
-        Logging.message(console.error, LoggingDebugLevel.EngineError, ...values)
+        Logging.message(console.error, LoggingLevel.EngineError, ...values)
     }
 
     static engineWarn(...values: any[]): void {
-        Logging.message(console.warn, LoggingDebugLevel.EngineWarn, ...values)
+        Logging.message(console.warn, LoggingLevel.EngineWarn, ...values)
     }
 
     static engineLog(...values: any[]): void {
-        Logging.message(console.log, LoggingDebugLevel.EngineLog, ...values)
+        Logging.message(console.log, LoggingLevel.EngineLog, ...values)
     }
 
     static engineSpam(...values: any[]): void {
-        Logging.message(console.log, LoggingDebugLevel.EngineSpam, ...values)
+        Logging.message(console.log, LoggingLevel.EngineSpam, ...values)
     }
 }